requirements.txt
discordpy
python-dotenv
